fix: Resolve MSVC type conversion warnings in 2D physics files
- Add static_cast<float>(M_PI) in RigidBody2D.cpp to fix C4305 double-to-float warnings
- Add static_cast<float>(M_PI) in CircleCollider2D.cpp to fix C4244 double-to-float warning
- Add static_cast<int>(m_objects.size()) in QuadTree.cpp to fix C4267 size_t-to-int warning

These changes resolve MSVC warnings being treated as errors (C2220) on Windows builds.

+++ src/Physics/2D/RigidBody2D.cpp
@@ -124,8 +124,8 @@ void RigidBody2D::IntegratePosition(float deltaTime) {
     if (!m_freezeRotation) {
         m_rotation += m_angularVelocity * deltaTime;
         
-        while (m_rotation > M_PI) m_rotation -= 2.0f * M_PI;
-        while (m_rotation < -M_PI) m_rotation += 2.0f * M_PI;
+        while (m_rotation > static_cast<float>(M_PI)) m_rotation -= 2.0f * static_cast<float>(M_PI);
+        while (m_rotation < -static_cast<float>(M_PI)) m_rotation += 2.0f * static_cast<float>(M_PI);
     }
 }
